rebranded generator to kontent branding - gradle project tests failing still but committing anyways for now.
