+++ .idea/.name
@@ -1,0 +1,1 @@
+kontent-generators-java
