--- .idea/modules/cloud-generator-core/cloud-generator-core.iml
@@ -1,13 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<module external.linked.project.id=":cloud-generator-core" external.linked.project.path="$MODULE_DIR$/../../../cloud-generator-core" external.root.project.path="$MODULE_DIR$/../../.." external.system.id="GRADLE" external.system.module.group="com.kenticocloud" external.system.module.version="1.1" type="JAVA_MODULE" version="4">
-  <component name="NewModuleRootManager" inherit-compiler-output="true">
-    <exclude-output />
-    <content url="file://$MODULE_DIR$/../../../cloud-generator-core">
-      <excludeFolder url="file://$MODULE_DIR$/../../../cloud-generator-core/.gradle" />
-      <excludeFolder url="file://$MODULE_DIR$/../../../cloud-generator-core/build" />
-      <excludeFolder url="file://$MODULE_DIR$/../../../cloud-generator-core/out" />
-    </content>
-    <orderEntry type="inheritedJdk" />
-    <orderEntry type="sourceFolder" forTests="false" />
-  </component>
-</module>
