+++ README.md
@@ -1,12 +1,12 @@
-# Kentico Cloud model generator utility for Java
+# Kontent model generator utility for Java
 
-[![Build Status](https://travis-ci.com/Kentico/cloud-generators-java.svg?branch=master)](https://travis-ci.com/Kentico/cloud-generators-java)
-[![Gradle Plugin](https://img.shields.io/static/v1?label=gradle&message=1.1&color=%3CCOLOR%3E?style=flat&logo=gradle&labelColor=02303A)](https://plugins.gradle.org/plugin/com.kenticocloud.generator)
+[![Build Status](https://travis-ci.com/Kentico/kontent-generators-java.svg?branch=master)](https://travis-ci.com/Kentico/kontent-generators-java)
+[![Gradle Plugin](https://img.shields.io/static/v1?label=gradle&message=1.1&color=%3CCOLOR%3E?style=flat&logo=gradle&labelColor=02303A)](https://plugins.gradle.org/plugin/com.github.kentico.kontent.generator)
 [![License: MIT](https://img.shields.io/badge/License-MIT-yellow.svg)](https://opensource.org/licenses/MIT)
 [![SonarQube](http://img.shields.io/badge/SonarQube-Results-blue.svg)](https://sonarcloud.io/organizations/kentico-github/projects)
-[![Stack Overflow](https://img.shields.io/badge/Stack%20Overflow-ASK%20NOW-FE7A16.svg?logo=stackoverflow&logoColor=white)](https://stackoverflow.com/tags/kentico-cloud)
+[![Stack Overflow](https://img.shields.io/badge/Stack%20Overflow-ASK%20NOW-FE7A16.svg?logo=stackoverflow&logoColor=white)](https://stackoverflow.com/tags/kontent)
 
-This utility generates strongly-typed models based on Content Types in a Kentico Cloud project. The models are supposed to be used together with the [Kentico Cloud Delivery SDK for Java](https://github.com/Kentico/delivery-sdk-java). Please read the [documentation](https://github.com/Kentico/delivery-sdk-java/wiki/Working-with-Strongly-Typed-Models-(aka-Code-First-Approach)#customizing-the-strong-type-binding-logic) to see all benefits of this approach.
+This utility generates strongly-typed models based on Content Types in a Kontent project. The models are supposed to be used together with the [Kontent Delivery SDK for Java](https://github.com/Kentico/delivery-sdk-java). Please read the [documentation](https://github.com/Kentico/delivery-sdk-java/wiki/Working-with-Strongly-Typed-Models-(aka-Code-First-Approach)#customizing-the-strong-type-binding-logic) to see all benefits of this approach.
 
 ## How to use
 
@@ -16,7 +16,7 @@ Apply the gradle plugin to your build file.
 
 ```groovy
 plugins {
-  id "com.kenticocloud.generator" version "1.1"
+  id "com.github.kentico.kontent.generator" version "1.1"
 }
 ```
 
@@ -28,11 +28,11 @@ buildscript {
 		mavenCentral()
 	}
 	dependencies {
-		classpath("com.kenticocloud:cloud-generators-gradle:1.1")
+		classpath("com.github.kentico:kontent-generators-gradle:1.1")
 	}
 }
 
-apply plugin: 'com.kenticocloud.generator'
+apply plugin: 'com.github.kentico.kontent.generator'
 ```
 
 Configure the plugin.
@@ -47,7 +47,7 @@ kenticoModel {
 }
 
 dependencies {
-    compile('com.kenticocloud:delivery-sdk-java:1.0.3')
+    compile('com.github.kentico:delivery-sdk-java:1.0.3')
 }
 ```
 
@@ -79,7 +79,7 @@ idea {
 ```
 
 ### Parameters
-- `projectId` - required - a GUID that can be found in [Kentico Cloud](https://app.kenticocloud.com) -> API keys -> Project ID
+- `projectId` - required - a GUID that can be found in [Kontent](https://app.kontent.ai) -> API keys -> Project ID
 - `packageName` - required - a name of the Java package to put your generated sources into
 - `outputDir` - required - a File object instance referencing an output folder path
 
@@ -88,18 +88,18 @@ idea {
 ```java
 package com.dancinggoat.models;
 
-import com.kenticocloud.delivery.Asset;
-import com.kenticocloud.delivery.ContentItem;
-import com.kenticocloud.delivery.ContentItemMapping;
-import com.kenticocloud.delivery.ElementMapping;
-import com.kenticocloud.delivery.System;
-import com.kenticocloud.delivery.Taxonomy;
+import com.github.kentico.delivery.Asset;
+import com.github.kentico.delivery.ContentItem;
+import com.github.kentico.delivery.ContentItemMapping;
+import com.github.kentico.delivery.ElementMapping;
+import com.github.kentico.delivery.System;
+import com.github.kentico.delivery.Taxonomy;
 import java.lang.String;
 import java.time.ZonedDateTime;
 import java.util.List;
 
 /**
- * This code was generated by a <a href="https://github.com/Kentico/cloud-generators-java">cloud-generators-java tool</a>
+ * This code was generated by a <a href="https://github.com/Kentico/kontent-generators-java">kontent-generators-java tool</a>
  *
  * Changes to this file may cause incorrect behavior and will be lost if the code is regenerated.
  * For further modifications of the class, create a separate file and extend this class.
@@ -218,11 +218,11 @@ public class Article {
 ```
 
 ## Feedback & Contributing
-Check out the [contributing](https://github.com/Kentico/cloud-generators-java/blob/master/CONTRIBUTING.md) page to see the best places to file issues, start discussions and begin contributing.
+Check out the [contributing](https://github.com/Kentico/kontent-generators-java/blob/master/CONTRIBUTING.md) page to see the best places to file issues, start discussions and begin contributing.
 
 ## Further information
 
-For more developer resources, visit the Kentico Cloud Developer Hub at <https://developer.kenticocloud.com>.
+For more developer resources, visit the Kontent Developer Hub at <https://developer.kontent.ai>.
 
 ### Building the sources
 
@@ -238,8 +238,8 @@ Ensure your `JAVA_HOME` environment is set.  Then build the project via the prov
 
 Optional:
 [JetBrains IntelliJ Idea](https://www.jetbrains.com/idea/) project files are included.  Open up the project and Import the Gradle project to sync up dependencies.
- * Note: The Gradle Test Kit in the cloud-generators-gradle module requires some extra steps to ensure it can access the 
+ * Note: The Gradle Test Kit in the kontent-generators-gradle module requires some extra steps to ensure it can access the 
  classpath of the plugin, which are handled in the gradle build.  When running the unit test from inside IntelliJ, you 
  may need to run the gradle build prior to ensure the latest build output is used if you are changing the main sources.
 
-![Analytics](https://kentico-ga-beacon.azurewebsites.net/api/UA-69014260-4/Kentico/cloud-generators-java/?pixel)
+![Analytics](https://kentico-ga-beacon.azurewebsites.net/api/UA-69014260-4/Kentico/kontent-generators-java/?pixel)
