--- cloud-generators-gradle/src/main/groovy/com/kenticocloud/generator/gradle/GenerateModelsPlugin.groovy
@@ -1,60 +1,0 @@
-/*
- * MIT License
- *
- * Copyright (c) 2017
- *
- * Permission is hereby granted, free of charge, to any person obtaining a copy
- * of this software and associated documentation files (the "Software"), to deal
- * in the Software without restriction, including without limitation the rights
- * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
- * copies of the Software, and to permit persons to whom the Software is
- * furnished to do so, subject to the following conditions:
- *
- * The above copyright notice and this permission notice shall be included in all
- * copies or substantial portions of the Software.
- *
- * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
- * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
- * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
- * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
- * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
- * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
- * SOFTWARE.
- */
-
-package com.kenticocloud.generator.gradle
-
-import org.gradle.api.Action
-import org.gradle.api.Plugin
-import org.gradle.api.Project
-import org.gradle.api.plugins.JavaPlugin
-import org.gradle.api.plugins.JavaPluginConvention
-import org.gradle.api.tasks.SourceSet
-
-import java.util.concurrent.Callable
-
-class GenerateModelsPlugin implements Plugin<Project> {
-
-    @Override
-    void apply(Project project) {
-        GenerateModelsExtension extension =
-                project.getExtensions().create("kenticoModel", GenerateModelsExtension.class)
-
-        project.getTasks().create("generateModels", GenerateModels.class)
-
-        project.getPlugins().withType(JavaPlugin.class, new Action<JavaPlugin>() {
-            @Override
-            void execute(JavaPlugin javaPlugin) {
-                JavaPluginConvention javaPluginConvention =
-                        project.getConvention().getPlugin(JavaPluginConvention.class)
-                SourceSet main = javaPluginConvention.getSourceSets().getByName(SourceSet.MAIN_SOURCE_SET_NAME)
-                main.getJava().srcDir(new Callable<File>(){
-                    @Override
-                    File call() throws Exception {
-                        return project.kenticoModel.outputDir
-                    }
-                })
-            }
-        })
-    }
-}
