--- cloud-generators-gradle/src/main/resources/META-INF/gradle-plugins/com.kenticocloud.generator.properties
@@ -1,1 +1,0 @@
-implementation-class=com.kenticocloud.generator.gradle.GenerateModelsPlugin
