+++ kontent-generator-core/src/test/resources/kentico/kontent/generator/SampleContentTypeList.json
