+++ kontent-generators-gradle/kontent-generators-gradle.iml
@@ -1,0 +1,60 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<module external.linked.project.id=":kontnt-generators-gradle" external.linked.project.path="$MODULE_DIR$" external.root.project.path="$MODULE_DIR$/.." external.system.id="GRADLE" type="JAVA_MODULE" version="4">
+  <component name="FacetManager">
+    <facet type="android-gradle" name="Android-Gradle">
+      <configuration>
+        <option name="GRADLE_PROJECT_PATH" value=":kontent-generators-gradle" />
+      </configuration>
+    </facet>
+    <facet type="java-gradle" name="Java-Gradle">
+      <configuration>
+        <option name="BUILD_FOLDER_PATH" value="$MODULE_DIR$/build" />
+        <option name="BUILDABLE" value="true" />
+      </configuration>
+    </facet>
+  </component>
+  <component name="NewModuleRootManager" LANGUAGE_LEVEL="JDK_1_8">
+    <output url="file://$MODULE_DIR$/build/classes/main" />
+    <output-test url="file://$MODULE_DIR$/build/classes/test" />
+    <exclude-output />
+    <content url="file://$MODULE_DIR$">
+      <sourceFolder url="file://$MODULE_DIR$/src/main/java" isTestSource="false" />
+      <sourceFolder url="file://$MODULE_DIR$/src/main/groovy" isTestSource="false" />
+      <sourceFolder url="file://$MODULE_DIR$/src/main/resources" type="java-resource" />
+      <sourceFolder url="file://$MODULE_DIR$/src/test/java" isTestSource="true" />
+      <sourceFolder url="file://$MODULE_DIR$/src/test/groovy" isTestSource="true" />
+      <sourceFolder url="file://$MODULE_DIR$/src/test/resources" type="java-test-resource" />
+      <excludeFolder url="file://$MODULE_DIR$/.gradle" />
+    </content>
+    <orderEntry type="inheritedJdk" />
+    <orderEntry type="sourceFolder" forTests="false" />
+    <orderEntry type="module" module-name="kontent-generator-core" exported="" />
+    <orderEntry type="library" exported="" name="Gradle: delivery-sdk-java-1.0.3" level="project" />
+    <orderEntry type="library" exported="" name="Gradle: javapoet-1.9.0" level="project" />
+    <orderEntry type="library" exported="" name="Gradle: guava-23.0" level="project" />
+    <orderEntry type="library" exported="" name="Gradle: jackson-core-2.8.9" level="project" />
+    <orderEntry type="library" exported="" name="Gradle: jackson-databind-2.8.9" level="project" />
+    <orderEntry type="library" exported="" name="Gradle: jackson-datatype-jsr310-2.4.0" level="project" />
+    <orderEntry type="library" exported="" name="Gradle: httpclient-4.5.3" level="project" />
+    <orderEntry type="library" exported="" name="Gradle: commons-beanutils-1.9.3" level="project" />
+    <orderEntry type="library" exported="" name="Gradle: fast-classpath-scanner-2.4.7" level="project" />
+    <orderEntry type="library" exported="" name="Gradle: jsr305-1.3.9" level="project" />
+    <orderEntry type="library" exported="" name="Gradle: error_prone_annotations-2.0.18" level="project" />
+    <orderEntry type="library" exported="" name="Gradle: j2objc-annotations-1.1" level="project" />
+    <orderEntry type="library" exported="" name="Gradle: animal-sniffer-annotations-1.14" level="project" />
+    <orderEntry type="library" exported="" name="Gradle: jackson-annotations-2.8.0" level="project" />
+    <orderEntry type="library" exported="" name="Gradle: httpcore-4.4.6" level="project" />
+    <orderEntry type="library" exported="" name="Gradle: commons-logging-1.2" level="project" />
+    <orderEntry type="library" exported="" name="Gradle: commons-codec-1.9" level="project" />
+    <orderEntry type="library" exported="" name="Gradle: commons-collections-3.2.2" level="project" />
+    <orderEntry type="library" exported="" name="Gradle: gradle-api-3.5-rc-2" level="project" />
+    <orderEntry type="library" exported="" name="Gradle: groovy-all-2.4.10" level="project" />
+    <orderEntry type="library" exported="" name="Gradle: gradle-installation-beacon-3.5" level="project" />
+    <orderEntry type="library" exported="" scope="TEST" name="Gradle: spock-core-1.0-groovy-2.4" level="project" />
+    <orderEntry type="library" exported="" scope="TEST" name="Gradle: junit-4.12" level="project" />
+    <orderEntry type="library" exported="" scope="TEST" name="Gradle: httpclient-4.5.3-tests" level="project" />
+    <orderEntry type="library" exported="" scope="TEST" name="Gradle: hamcrest-core-1.3" level="project" />
+    <orderEntry type="library" exported="" scope="TEST" name="Gradle: C__Users_timothyf_Documents_cloud_generators_java_cloud_generators_gradle_build_pluginUnderTestMetadata" level="project" />
+    <orderEntry type="library" exported="" scope="TEST" name="Gradle: gradle-test-kit-3.5-rc-2" level="project" />
+  </component>
+</module>
