+++ kontent-generators-gradle/src/main/resources/META-INF/gradle-plugins/kentico.kontent.generator.properties
@@ -1,0 +1,1 @@
+implementation-class=kentico.kontent.generator.gradle.GenerateModelsPlugin
