+++ kontent-generators-gradle/src/test/resources/kentico/generator/gradle/SampleContentTypeList.json
